src/main/java/com/oocl/web/sampleWebApp/domain/ParkingLot.java
package com.oocl.web.sampleWebApp.domain;

import org.hibernate.validator.constraints.Range;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;


@Entity
@Table(name = "parking_lot")
public class ParkingLot {
    @Id
    private Long Id;

    @Column(name = "parkingLotID", unique = true, nullable = false)
    private String parkingLotId;

    @Range(min = 1, max = 100)
    private int capacity;
    public ParkingLot(String lot, int i) {}

    public int getCapacity() {
        return capacity;
    }
    public void setCapacity(int capacity) {
        this.capacity = capacity;
    }

    public Long getId() {
        return Id;
    }

    public void setId(Long id) {
        Id = id;
    }

    public String getParkingLotId() {
        return parkingLotId;
    }

    public void setParkingLotId(String parkingLotId) {
        this.parkingLotId = parkingLotId;
    }
}
